+++ R/easyRasch.R
@@ -5141,6 +5141,56 @@ RIbootRestscore <- function(dat, iterations = 200, samplesize = 600, cpu = 4,
 }
 
 
+#' Bootstrapped Likelihood Ratio Test
+#'
+#' Non-parametric bootstrap use of `iarm::clr_tests()`.
+#'
+#' @param dat A dataframe with response data
+#' @param iterations How many bootstrap samples to run
+#' @param samplesize How large sample to use in each bootstrap
+#' @param cpu How many CPU's to use
+#' @export
+RIbootCLR <- function(dat, iterations = 250, samplesize = 500, cpu = 4) {
+
+  if(min(as.matrix(dat), na.rm = T) > 0) {
+    stop("The lowest response category needs to coded as 0. Please recode your data.")
+  } else if (samplesize > nrow(dat)) {
+    stop(paste0("`samplesize` (",samplesize,") cannot be larger than the number of rows in your data (",
+                nrow(dat),")."))
+  } else if(max(as.matrix(dat), na.rm = T) == 1) {
+    model <- "RM"
+    relative_item_avg_locations <- item_avg_locations - person_avg_locations
+  } else if(max(as.matrix(dat), na.rm = T) > 1) {
+    model <- "PCM"
+  }
+
+  require(doParallel)
+  registerDoParallel(cores = cpu)
+
+  fit <- data.frame()
+  fit <- foreach(i = 1:iterations, .combine = rbind) %dopar% {
+
+    data <- dat[sample(1:nrow(dat), samplesize, replace = TRUE), ]
+
+    if (model == "PCM") {
+      lrt_out <- clr_tests(data, model = "PCM")[3]
+    } else if (model == "RM") {
+      lrt_out <- clr_tests(data, model = "RM")[3]
+    }
+
+    as.numeric(lrt_out)
+
+  }
+
+  fit %>%
+    as.data.frame() %>%
+    set_names("pvalue") %>%
+    mutate(Result = ifelse(pvalue < .05, "Statistically significant", "Not statistically significant")) %>%
+    count(Result) %>%
+    mutate(Percent = round(n*100/sum(n),1)) %>%
+    knitr::kable()
+}
+
 
 #' Temporary fix for upstream bug in `iarm::person_estimates()`
 #'